serial for postgres

+++ src/main/resources/db/migration/V1__init.sql
@@ -1,5 +1,5 @@
 CREATE TABLE PERSON (
-	id int8 not null primary key,
+	id serial not null primary key,
 	first_name varchar(255) not null,
 	last_name varchar(255) not null
 );
